Make the int-to-float and size_t-to-streamsize conversions explicit
bugprone-narrowing-conversions, three sites. The conversions are all
intended, so the change is to say so in the code rather than to alter
any arithmetic. Verified: clang-tidy reports zero findings afterwards.

CreateSphereGeometry() multiplies a loop counter by a float step
(stack * stackStep, slice * sliceStep). Neighbouring lines do the same
kind of conversion and are not reported, which is not an inconsistency
in the check: it folds constants and stays quiet when the value is
exactly representable as a float. sliceCount is a const int of 48, the
loop counters are not constants. Measured on a scratch file — a const
int of 16777217, which a float cannot represent exactly, is reported
with a different message that names the value.

LoadShaderModule() passes a size_t to std::istream::read(), whose
second parameter is std::streamsize (signed). The value comes from
tellg() on a .spv file, so it cannot approach the point where the
conversion would be implementation-defined. <ios> is already included,
so std::streamsize needs no new include.

Also drops the C-style (float) cast on kPi, which was already a float.
Leaving a cast that does nothing next to a cast that does something
would make the line harder to read, not easier.

The alternative was to set
bugprone-narrowing-conversions.WarnOnIntegerToFloatingPointNarrowingConversion
to false, which is how the option is described in .clang-tidy for the
day it gets noisy. Three sites in the whole project is not that day, and
the explicit casts document intent that the option would only hide.

+++ 02_simplecube/simplecube_app.cpp
@@ -145,10 +145,10 @@ void SimpleCubeApp::CreateSphereGeometry() {
 
     std::vector<Vertex> vertices;
     for (int stack = 0; stack <= stackCount; ++stack) {
-        auto stackAngle = (float)kPi / 2 - stack * stackStep;
+        auto stackAngle = kPi / 2 - static_cast<float>(stack) * stackStep;
 
         for (int slice = 0; slice <= sliceCount; ++slice) {
-            auto sliceAngle = slice * sliceStep;
+            auto sliceAngle = static_cast<float>(slice) * sliceStep;
 
             auto x = std::cos(stackAngle) * std::cos(sliceAngle);
             auto y = std::sin(stackAngle);
